requirements.txt
nose
rednose
sphinx
numpy
pyopengl==3.1.1a1